PDJB-766: Use base pages for common page objects
