+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/PropertyRegistrationJourneyTests.kt
@@ -432,7 +432,6 @@ class PropertyRegistrationJourneyTests : IntegrationTestWithMutableData("data-lo
 
         // Has EPC - render page
         assertThat(hasEpcPage.form.sectionHeader).containsText("Section 1 of 2 — Register your property details")
-        assertThat(hasEpcPage.heading).containsText("Do you have an EPC for this property?")
         hasEpcPage.submitHasEpc()
         val findYourEpcPage = assertPageIs(page, FindYourEpcFormPagePropertyRegistration::class)
 
@@ -633,7 +632,6 @@ class PropertyRegistrationJourneyTests : IntegrationTestWithMutableData("data-lo
 
         // Has EPC - render page
         assertThat(hasEpcPage.form.sectionHeader).containsText("Section 1 of 2 — Register your property details")
-        assertThat(hasEpcPage.heading).containsText("Do you have an EPC for this property?")
         hasEpcPage.submitHasNoEpc()
         val isEpcRequiredPage = assertPageIs(page, IsEpcRequiredFormPagePropertyRegistration::class)
 
@@ -735,7 +733,6 @@ class PropertyRegistrationJourneyTests : IntegrationTestWithMutableData("data-lo
 
         // Has EPC - render page
         assertThat(hasEpcPage.form.sectionHeader).containsText("Section 1 of 2 — Register your property details")
-        assertThat(hasEpcPage.heading).containsText("Do you have an EPC for this property?")
         hasEpcPage.submitProvideThisLater()
         val provideEpcLaterPage = assertPageIs(page, ProvideEpcLaterFormPagePropertyRegistration::class)
 
@@ -810,7 +807,6 @@ class PropertyRegistrationJourneyTests : IntegrationTestWithMutableData("data-lo
 
         // Has EPC - render page
         assertThat(hasEpcPage.form.sectionHeader).containsText("Section 1 of 2 — Register your property details")
-        assertThat(hasEpcPage.heading).containsText("Do you have an EPC for this property?")
         hasEpcPage.submitProvideThisLater()
         val provideEpcLaterPage = assertPageIs(page, ProvideEpcLaterFormPagePropertyRegistration::class)
 
@@ -879,7 +875,6 @@ class PropertyRegistrationJourneyTests : IntegrationTestWithMutableData("data-lo
 
         // Has EPC - render page
         assertThat(hasEpcPage.form.sectionHeader).containsText("Section 1 of 2 — Register your property details")
-        assertThat(hasEpcPage.heading).containsText("Do you have an EPC for this property?")
         hasEpcPage.submitHasNoEpc()
         val isEpcRequiredPage = assertPageIs(page, IsEpcRequiredFormPagePropertyRegistration::class)
 
@@ -1129,7 +1124,6 @@ class PropertyRegistrationJourneyTests : IntegrationTestWithMutableData("data-lo
 
         // Has EPC - render page
         assertThat(hasEpcPage.form.sectionHeader).containsText("Section 1 of 2 — Register your property details")
-        assertThat(hasEpcPage.heading).containsText("Do you have an EPC for this property?")
         hasEpcPage.submitHasEpc()
         val findYourEpcPage = assertPageIs(page, FindYourEpcFormPagePropertyRegistration::class)
