+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/propertyRegistrationJourneyPages/HasEpcFormPagePropertyRegistration.kt
@@ -1,42 +1,10 @@
 package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages
 
 import com.microsoft.playwright.Page
-import uk.gov.communities.prsdb.webapp.constants.CONTINUE_BUTTON_ACTION_NAME
-import uk.gov.communities.prsdb.webapp.constants.PROVIDE_THIS_LATER_BUTTON_ACTION_NAME
 import uk.gov.communities.prsdb.webapp.controllers.RegisterPropertyController
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.FormWithSectionHeader
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.Heading
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.Radios
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.HasEpcFormBasePage
 import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.HasEpcStep
 
 class HasEpcFormPagePropertyRegistration(
     page: Page,
-) : BasePage(page, "${RegisterPropertyController.PROPERTY_REGISTRATION_ROUTE}/${HasEpcStep.ROUTE_SEGMENT}") {
-    val heading = Heading(page.locator("h1"))
-    val form = HasEpcForm(page)
-
-    fun submitHasEpc() {
-        form.hasCertRadios.selectValue("true")
-        form.submitPrimaryButton()
-    }
-
-    fun submitHasNoEpc() {
-        form.hasCertRadios.selectValue("false")
-        form.submitPrimaryButton()
-    }
-
-    fun submitProvideThisLater() {
-        form.submitSecondaryButton()
-    }
-
-    class HasEpcForm(
-        page: Page,
-    ) : FormWithSectionHeader(page) {
-        val hasCertRadios = Radios(locator)
-
-        fun submitPrimaryButton(buttonAction: String = CONTINUE_BUTTON_ACTION_NAME) = submitSelectedButton(buttonAction)
-
-        fun submitSecondaryButton(buttonAction: String = PROVIDE_THIS_LATER_BUTTON_ACTION_NAME) = submitSelectedButton(buttonAction)
-    }
-}
+) : HasEpcFormBasePage(page, "${RegisterPropertyController.PROPERTY_REGISTRATION_ROUTE}/${HasEpcStep.ROUTE_SEGMENT}")
